trait based itinerary modifier interface
