+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod infection_importation;
 pub mod infection_propagation_loop;
 pub mod infectiousness_manager;
 pub mod itinerary_manager;
+pub mod itinerary_modifiers;
 pub mod model;
 pub mod natural_history_parameter_manager;
 pub mod parameters;
